requirements.txt
asgiref==3.4.1
attrs==22.1.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
crashtest==0.3.1
distlib==0.3.6
dj-database-url==1.0.0
Django==4.0.8
django-cors-headers==3.13.0
djangorestframework==3.14.0
dulwich==0.20.46
filelock==3.8.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
jaraco.classes==3.2.3
jsonschema==4.16.0
keyring==23.9.3
lockfile==0.12.2
more-itertools==9.0.0
msgpack==1.0.4
packaging==21.3
pexpect==4.8.0
pkginfo==1.8.3
platformdirs==2.5.2
psycopg==3.1.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pycparser==2.21
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.5
requests==2.28.1
requests-toolbelt==0.9.1
shellingham==1.5.0
six==1.16.0
sqlparse==0.4.3
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.16.6
webencodings==0.5.1
xattr==0.9.9
